requirements.txt
attrs==19.3.0
Babel==2.8.0
brother-ql==0.9.4
click==7.1.1
future==0.18.2
packbits==0.6
Pillow==7.0.0
pkg-resources==0.0.0
pytz==2019.3
pyusb==1.0.2
qrcode==6.1
six==1.14.0
tkcalendar==1.6.1
